refactor: move workerset verifier logic to its own file
