+++ ampd/src/handlers/solana_verify_worker_set.rs
@@ -18,7 +18,7 @@ use crate::event_processor::EventHandler;
 use crate::handlers::errors::Error;
 use crate::queue::queued_broadcaster::BroadcasterClient;
 use crate::solana::json_rpc::SolanaClient;
-use crate::solana::verifier::{parse_gateway_event, verify_worker_set, GatewayEvent};
+use crate::solana::ws_verifier::{parse_gateway_event, verify_worker_set, GatewayEvent};
 use crate::types::{TMAddress, U256};
 
 type Result<T> = error_stack::Result<T, Error>;
