+++ ampd/src/solana/mod.rs
@@ -1,4 +1,4 @@
 pub mod json_rpc;
 mod pub_key_wrapper;
-pub mod verifier;
+pub mod ws_verifier;
